refactor: create distinct process route tests file
